First step towards the new Ranking Logic
The new ranking logic is based on RankingComponents, which form chains
to determine the sort order. RankingComponentIndex builds those chains
from an array of keywords (strings). At the root of every ComponentChain
is an IDRankingComponent to ensure unambiguous ranking orders.

Different sets of RankingComponents can be ordered in any way imaginable
For now, they include:

* GamesRankingComponent: number of games
* IDRankingComponent: player id (for unambiguous rankings)
* PointsRankingComponent: won points (small points)
* SaldoRankingComponent: saldo points (point difference)
* WinsRankingComponent: number of won games

A dependency system for the data components (this.wins, this.wonpoints,
this.lostpoints, ...) is required.

A small additional change, create-core-class.sh lets me specify the base
class in an optional second argument