adding VectorModel.add and VectorModel.mult
VectorModel.prototype.add(a,b): element-wise add the vectors a and b and
store the results in this. b defaults to this, so this function can be
used to add a vector to this.

VectorModel.prototype.mult(a,b) works the same way, but with
multiplication. This supersedes setProduct, which is removed