write-sprite.sh: Reading image names from html
When rebuilding the sprite, the required images are now read from the
target html files (*.html in the target directories) instead of reading
the file system contents. This avoids loading unnecessary images and
further reduces the file size of the sprite.

In addition, it allows for a central image directory.