Tests can now be built using `make test`
The TestEnv pattern was inconvenient, so it is changed to a TestMain
function for every view test. Works just as well, but the config can be
inside a shared config file.

test/index.html is auto-generated during the build process and links to
the Unit Tests and the View Tests.

test/scripts/build.js is added for optimizing the scripts during the
build process.

`make test` works.