Tab_Debug: re-enabling player autoregistration
The player registration from debugging tab is re-enabled in this
commit. It had been disabled due to the big rewrite of Tab_Teams.

As a side effect, player registration is allowed after the registration
has been closed, since a closed registration is purely visual. This
decision is deliberate, since it helps in debugging the running software
by adding teams during a running tournament, which is impossible in a
production release.