Tab_Games: focus #new and toast after last game
About half a year ago, the multi-tournament feature broke the
"lastgame?"-check, which is used to focus the #new tab and emit a Toast
when the #games tab is about to become hidden. This feature has been
restored with a cheap hack, which is supposed to be implemented properly
during the big rewrite.