Manual link disabled; NEWS updated from commit log
Manual links: The manual is not readable at the moment, so I disabled
the manual links prior to release. Since it's not likely to be written
in the near future, since Tuvero should be self-explanatory and since no
user ever read a manual, this should not make too much of a difference.

NEWS: The NEWS file was updated from the commit log.