KOTournament: support for third, fifth, etc. place
Using a roundid for every game and player, the best achievable place for
every player can be determined, along with the next player in the
respective sub-tree inside the KO tournament. That way, every place can
be played off.

winner round is 0, third place round is 1, fifth place round is 2, etc.