adding and testing MatchReferenceListModel
MatchReferenceListModel is a readonly ListModel instance, which is
tightly bound to a list of MatchModel instances. It mirrors the
MatchModel list, but translates the team ids to external team ids,
according to a teamid list, as it's used in TournamentModel.

This enables the transparent display and finish of matches without
further knowledge of the TournamentModel implementation.